+++ cognite/neat/_session/_prepare.py
@@ -38,17 +38,70 @@ class DataModelPrepareAPI:
                 )
             )
 
-    def to_extension(self, new_data_model_id: DataModelIdentifier, org_name: str | None = None) -> None:
-        """Uses the current data model as a basis to extend from.
+    def to_enterprise(
+        self,
+        data_model_id: DataModelIdentifier,
+        org_name: str = "My",
+        dummy_property: str = "GUID",
+    ) -> None:
+        """Uses the current data model as a basis to create enterprise data model
 
         Args:
-            new_data_model_id: The new data model that is extending the current data model.
-            org_name: Organization name to use for the views in the new data model. This is required if you are
-                creating an extension from a Cognite Data Model.
+            data_model_id: The enterprise data model id that is being created
+            org_name: Organization name to use for the views in the enterprise data model.
+            dummy_property: The dummy property to use as placeholder for the views in the new data model.
+
+        !!! note "Enterprise Data Model Creation"
+            Always create an enterprise data model from a Cognite Data Model as this will
+            assure all the Cognite Data Fusion applications to run smoothly, such as
+            - Search
+            - Atlas AI
+            - ...
 
         """
         if dms := self._state.last_verified_dms_rules:
-            output = ToExtension(new_data_model_id, org_name).transform(dms)
+            output = ToExtension(
+                new_model_id=data_model_id,
+                org_name=org_name,
+                type_="enterprise",
+                dummy_property=dummy_property,
+            ).transform(dms)
+            self._state.verified_rules.append(output.rules)
+
+    def to_solution(
+        self,
+        data_model_id: DataModelIdentifier,
+        org_name: str = "My",
+        mode: Literal["read", "write"] = "read",
+        dummy_property: str = "dummy",
+    ) -> None:
+        """Uses the current data model as a basis to create solution data model
+
+        Args:
+            data_model_id: The solution data model id that is being created.
+            org_name: Organization name to use for the views in the new data model.
+            mode: The mode of the solution data model. Can be either "read" or "write".
+            dummy_property: The dummy property to use as placeholder for the views in the new data model.
+
+        !!! note "Solution Data Model Mode"
+            The read-only solution model will only be able to read from the existing containers
+            from the enterprise data model, therefore the solution data model will not have
+            containers in the solution data model space. Meaning the solution data model views
+            will be read-only.
+
+            The write mode will have additional containers in the solution data model space,
+            allowing in addition to reading through the solution model views, also writing to
+            the containers in the solution data model space.
+
+        """
+        if dms := self._state.last_verified_dms_rules:
+            output = ToExtension(
+                new_model_id=data_model_id,
+                org_name=org_name,
+                type_="solution",
+                mode=mode,
+                dummy_property=dummy_property,
+            ).transform(dms)
             self._state.verified_rules.append(output.rules)
 
     def reduce(self, drop: Collection[Literal["3D", "Annotation", "BaseViews"] | str]) -> None: