+++ cognite/neat/_session/_state.py
@@ -1,6 +1,7 @@
 from dataclasses import dataclass, field
 from typing import Literal, cast
 
+from cognite.neat._issues import IssueList
 from cognite.neat._rules._shared import ReadRules, VerifiedRules
 from cognite.neat._rules.models.dms._rules import DMSRules
 from cognite.neat._rules.models.information._rules import InformationRules
@@ -15,6 +16,7 @@ class SessionState:
     store_type: Literal["memory", "oxigraph"]
     input_rules: list[ReadRules] = field(default_factory=list)
     verified_rules: list[VerifiedRules] = field(default_factory=list)
+    issue_lists: list[IssueList] = field(default_factory=list)
     _store: NeatGraphStore | None = field(init=False, default=None)
 
     @property
@@ -49,21 +51,39 @@ class SessionState:
         return self.verified_rules[-1]
 
     @property
-    def last_verified_dms_rules(self) -> DMSRules | None:
+    def last_verified_dms_rules(self) -> DMSRules:
         if self.verified_rules:
             for rules in self.verified_rules[::-1]:
                 if isinstance(rules, DMSRules):
                     return rules
-        return None
+
+        raise NeatSessionError(
+            'No verified DMS data model. Try using  [bold].convert("DMS")[/bold]'
+            " to convert verified information model to verified DMS model."
+        )
 
     @property
-    def last_verified_information_rules(self) -> InformationRules | None:
+    def last_verified_information_rules(self) -> InformationRules:
         if self.verified_rules:
             for rules in self.verified_rules[::-1]:
                 if isinstance(rules, InformationRules):
                     return rules
-        return None
+
+        raise NeatSessionError(
+            "No verified information data model. Try using  [bold].verify()[/bold]"
+            " to convert unverified information model to verified information model."
+        )
 
     @property
     def has_store(self) -> bool:
         return self._store is not None
+
+    @property
+    def has_verified_rules(self) -> bool:
+        return bool(self.verified_rules)
+
+    @property
+    def last_issues(self) -> IssueList:
+        if not self.issue_lists:
+            raise NeatSessionError("No issues available. Try using [bold].verify()[/bold] to verify a data model.")
+        return self.issue_lists[-1]