+++ docs/CHANGELOG.md
@@ -15,6 +15,10 @@ Changes are grouped as follows:
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
+## TBD
+### Improved
+- Handling of CDM extension
+
 ## [0.95.0] - 29-10-**2024**
 ### Fixed
 - `NeatSession` subcommands no longer gives traceback for `NeatSessionError` exceptions, instead it