+++ docs/CHANGELOG.md
@@ -18,6 +18,7 @@ Changes are grouped as follows:
 ## TBD
 ### Added
 - `neat.prepare.reduce` now support dropping individual views from a `Cognite` model.
+- `neat.set.data_model_id` a convenience method to set the data model id in a `NeatSession`.
 
 ## [0.95.0] - 29-10-**2024**
 ### Fixed