+++ docs/CHANGELOG.md
@@ -18,10 +18,13 @@ Changes are grouped as follows:
 ## TBD
 ### Improved
 - Handling of CDM extension
+- Switched from Cytoscape to PyVis for data model and instances visualization
 ### Added
 - `neat.prepare.reduce` now support dropping individual views from a `Cognite` model.
 - `neat.set.data_model_id` a convenience method to set the data model id in a `NeatSession`.
 - `neat.version` returns the version of the `neat` package.
+- `neat.prepare.to_enterprise` prepares template for creation of an enterprise model in `Cognite Data Fusion`.
+- `neat.prepare.to_solution` prepares template for creation of a solution model in `Cognite Data Fusion`.
 
 
 ## [0.95.0] - 29-10-**2024**